tsmc_hackthon/Report_agent/requirements.txt
# Core LLM Framework
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0

# LangGraph for Multi-Agent Orchestration
langgraph>=0.0.40

# Google Gemini AI
google-generativeai>=0.3.0

# Environment Management
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# HTTP Requests
httpx>=0.25.0

# Logging
logging

# JSON handling (built-in)
json

# Type hints (built-in)
typing

# Date/Time handling (built-in)
datetime
